pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-reliability-framework"
version = "4.0.1+oss"
description = "Agentic Reliability Framework - OSS Edition"
readme = "README.md"
authors = [
    {name = "Juan Petter", email = "petter2025us@outlook.com"}
]
license = {text = "Apache-2.0"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "numpy",
    "pyyaml",
    "circuitbreaker>=2.0.0",
    "pymc>=5.0.0",
    "arviz>=0.15.0",
    "anthropic>=0.8.1",
    "pandas>=2.0.0",
    "scikit-learn>=1.0.0",
    "faiss-cpu>=1.7.0",               # Added for memory
    "sentence-transformers>=2.2.0",    # Added for embeddings
]

[project.urls]
Homepage = "https://github.com/petter2025us/agentic-reliability-framework"

[tool.setuptools.packages.find]
include = ["agentic_reliability_framework*"]
